edited after feedback
