+++ src/components/restaurantContent/component.jsx
@@ -1,51 +1,26 @@
 import { useState } from 'react';
 import { restaurants } from '../../../data/mock'
 import './restaurantContent.css'
-import { Dish } from '../dish/component';
+import { Restaurant } from '../restaurant/component';
 
 export const RestaurantContent = () => {
-    const [restaurantId, getRestaurant] = useState(restaurants[0].id);
+    const [activeRestaurantId, setRestaurant] = useState(restaurants[0].id);
 
     const showRestaurant = (restaurantId) => {
-        getRestaurant(restaurantId);
+        setRestaurant(restaurantId);
     };
 
     return (
         <>
             <div className='button-container'>
                 {
-                    restaurants.map((r) => 
+                    restaurants.map((restaurant_) => 
                         <button onClick={(e) => {
-                            showRestaurant(r.id);}}>{r.name}</button>
+                            showRestaurant(restaurant_.id);}}>{restaurant_.name}</button>
                     )
                 }
             </div>
-           { restaurants.map((restaurant) => 
-                restaurant.id == restaurantId 
-                    ? (<div className='restaurant-container'>
-                        <h2>{restaurant.name}</h2>
-                        <h3>Menu:</h3>
-                        <ul>
-                            {restaurant.menu.map((dish) => 
-                                <Dish id={dish.id} name={dish.name} price={dish.price} ingredients={dish.ingredients} />
-                            )}
-                        </ul>
-                        <h3>Reviews:</h3>
-                        <ul>
-                            {restaurant.reviews && restaurant.reviews.length 
-                                ? restaurant.reviews.map((review) => 
-                                    <li>
-                                        <h4>{review.user}, {review.rating}★</h4>
-                                        <p>{review.text}</p>
-                                    </li>
-                                )
-                                : <h4>No reviews</h4>
-                            }
-                        </ul>
-                     </div>)
-                    : <></>
-           
-            )}
+            <Restaurant restaurantId={activeRestaurantId}/>
         </>
     );
 };
